LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoFrame-copyTo.any-expected.txt


FAIL Test closed frame. assert_throws_dom: allocationSize() function "() => frame.allocationSize()" threw object "TypeError: Member VideoFrameCopyToOptions.rect is required and must be an instance of DOMRectInit" that is not a DOMException InvalidStateError: property "code" is equal to undefined, expected 11
FAIL Test copying I420 frame to a non-shared ArrayBuffer promise_test: Unhandled rejection with value: object "TypeError: Member VideoFrameCopyToOptions.rect is required and must be an instance of DOMRectInit"
FAIL Test copying I420 frame to a non-shared ArrayBufferView promise_test: Unhandled rejection with value: object "TypeError: Member VideoFrameCopyToOptions.rect is required and must be an instance of DOMRectInit"
FAIL Test RGBA frame. promise_test: Unhandled rejection with value: object "TypeError: Member VideoFrameCopyToOptions.rect is required and must be an instance of DOMRectInit"
PASS Test undersized buffer.
PASS Test incorrect plane count.
FAIL Test stride and offset work. promise_test: Unhandled rejection with value: object "TypeError: Member VideoFrameCopyToOptions.rect is required and must be an instance of DOMRectInit"
FAIL Test stride and offset with padding. promise_test: Unhandled rejection with value: object "TypeError: Member VideoFrameCopyToOptions.rect is required and must be an instance of DOMRectInit"
PASS Test invalid stride.
PASS Test address overflow.
FAIL Test codedRect. assert_equals: allocationSize() expected 12 but got 0
FAIL Test empty rect. assert_throws_js: function "() => frame.allocationSize(options)" did not throw
FAIL Test unaligned rect. assert_throws_js: function "() => frame.allocationSize(options)" did not throw
FAIL Test left crop. assert_equals: allocationSize() expected 6 but got 0
FAIL Test invalid rect. assert_throws_js: function "() => frame.allocationSize(options)" did not throw
